+++ src/Ransac.cpp
@@ -15,7 +15,7 @@
 // =====================================================
 // Process Ransac on a pointcloud cell
 // =====================================================
-void RansacAlgorithm::FitPlane(int t, vector<pcl::PointXYZI>& data, vector<float>& bestFit)
+void RansacAlgorithm::FitPlane(int t, vector<pcl::PointXYZI>& data, vector<double>& bestFit)
 {
     int best_inliers = 0;
     random_device dev;
@@ -36,7 +36,7 @@ void RansacAlgorithm::FitPlane(int t, vector<pcl::PointXYZI>& data, vector<float
             while (indx_3 == indx_1 || indx_3 == indx_2) indx_3 = dist_data(rng);
 
             // Compute plane equation based on the sample taken
-            vector<float> plane = RansacAlgorithm::PlaneEquation(data[indx_1],data[indx_2],data[indx_3]);
+            vector<double> plane = RansacAlgorithm::PlaneEquation(data[indx_1],data[indx_2],data[indx_3]);
 
 
             // Evaluate inliners
@@ -55,11 +55,11 @@ void RansacAlgorithm::FitPlane(int t, vector<pcl::PointXYZI>& data, vector<float
 };
 
 
-vector<float> RansacAlgorithm::PlaneEquation(const pcl::PointXYZI& p1, const pcl::PointXYZI& p2, const pcl::PointXYZI& p3)
+vector<double> RansacAlgorithm::PlaneEquation(const pcl::PointXYZI& p1, const pcl::PointXYZI& p2, const pcl::PointXYZI& p3)
 {
     // Calculate two vectors that lies on the plane we wish to calculate
-    vector<float> v1({p2.x-p1.x, p2.y-p1.y, p2.z-p1.z});
-    vector<float> v2({p3.x-p1.x, p3.y-p1.y, p3.z-p1.z}); 
+    vector<double> v1({p2.x-p1.x, p2.y-p1.y, p2.z-p1.z});
+    vector<double> v2({p3.x-p1.x, p3.y-p1.y, p3.z-p1.z}); 
 
     // Calculate the cross product between those two vectors. -> Gives us a normal vector to the plane (aka A,B and C coefficients of the plane equation).
     //           |  A   B   C  |
@@ -70,13 +70,13 @@ vector<float> RansacAlgorithm::PlaneEquation(const pcl::PointXYZI& p1, const pcl
     // B = V2x * V1z - V1x * V2z
     // C = V1x * V2y - V2x * V1y
 
-    float A = v1[1]*v2[2] - v2[1]*v1[2];
-    float B = v2[0]*v1[2] - v1[0]*v2[2];
-    float C = v1[0]*v2[1] - v2[0]*v1[1];
+    double A = v1[1]*v2[2] - v2[1]*v1[2];
+    double B = v2[0]*v1[2] - v1[0]*v2[2];
+    double C = v1[0]*v2[1] - v2[0]*v1[1];
 
     // Calculate the coefficient D using the plane equation Ax + By + Cz = D and one of the input points
 
-    float D = A*p1.x + B*p1.y + C*p1.z;
+    double D = A*p1.x + B*p1.y + C*p1.z;
 
     return {A,B,C,D};
 };
@@ -90,23 +90,23 @@ vector<float> RansacAlgorithm::PlaneEquation(const pcl::PointXYZI& p1, const pcl
 // TOOLS
 // =====================================================
 
-float RansacAlgorithm::CalculateDistanceFromPlane(const pcl::PointXYZI& point, const vector<float>& plane_eq)
+double RansacAlgorithm::CalculateDistanceFromPlane(const pcl::PointXYZI& point, const vector<double>& plane_eq)
 {
-    float up = abs(plane_eq[0] * point.x + plane_eq[1] * point.y + plane_eq[2] * point.z + plane_eq[3]);
-    float down = sqrt(plane_eq[0]*plane_eq[0] + plane_eq[1]*plane_eq[1] + plane_eq[2]*plane_eq[2]);
+    double up = abs(plane_eq[0] * point.x + plane_eq[1] * point.y + plane_eq[2] * point.z + plane_eq[3]);
+    double down = sqrt(plane_eq[0]*plane_eq[0] + plane_eq[1]*plane_eq[1] + plane_eq[2]*plane_eq[2]);
     return (up/down);
 };
 
 
-int RansacAlgorithm::CountInliers(vector<pcl::PointXYZI>& data, vector<float>& plane_eq, int t)
+int RansacAlgorithm::CountInliers(vector<pcl::PointXYZI>& data, vector<double>& plane_eq, int t)
 {
     int inliners_number = 0;
-    vector<float> temp_distances;                       // Temp distance vector
+    vector<double> temp_distances;                       // Temp distance vector
     RansacAlgorithm::temp_distances = temp_distances;
 
     for(long unsigned int i=0; i<data.size(); i++)
     {
-        float distance = RansacAlgorithm::CalculateDistanceFromPlane(data[i], plane_eq);
+        double distance = RansacAlgorithm::CalculateDistanceFromPlane(data[i], plane_eq);
         RansacAlgorithm::temp_distances.push_back(distance);
 
         if(distance < t)
